video generation service/veo-api-service/requirements.txt
# ============================================================================
# FILE: veo-api-service/requirements.txt
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
google-auth==2.23.4